+++ internal/controllers/gateway/udproute_controller.go
@@ -351,7 +351,10 @@ func (r *UDPRouteReconciler) Reconcile(ctx context.Context, req ctrl.Request) (c
 
 		if configurationStatus == k8sobj.ConfigurationStatusFailed {
 			debug(log, udproute, "tcproute configuration failed")
-			statusUpdated, err := r.ensureParentsProgrammedCondition(ctx, udproute, gateways, metav1.ConditionFalse, ConditionReasonTranslationError, "")
+			statusUpdated, err := ensureParentsProgrammedCondition(ctx, r.Status(), udproute, udproute.Status.Parents, gateways, metav1.Condition{
+				Status: metav1.ConditionFalse,
+				Reason: string(ConditionReasonTranslationError),
+			})
 			if err != nil {
 				// don't proceed until the statuses can be updated appropriately
 				debug(log, udproute, "failed to update programmed condition")
@@ -360,7 +363,10 @@ func (r *UDPRouteReconciler) Reconcile(ctx context.Context, req ctrl.Request) (c
 			return ctrl.Result{Requeue: !statusUpdated}, nil
 		}
 
-		statusUpdated, err := r.ensureParentsProgrammedCondition(ctx, udproute, gateways, metav1.ConditionTrue, ConditionReasonConfiguredInGateway, "")
+		statusUpdated, err := ensureParentsProgrammedCondition(ctx, r.Status(), udproute, udproute.Status.Parents, gateways, metav1.Condition{
+			Status: metav1.ConditionTrue,
+			Reason: string(ConditionReasonConfiguredInGateway),
+		})
 		if err != nil {
 			// don't proceed until the statuses can be updated appropriately
 			debug(log, udproute, "failed to update programmed condition")
@@ -497,62 +503,4 @@ func (r *UDPRouteReconciler) ensureGatewayReferenceStatusRemoved(ctx context.Con
 
 	// the status needed an update and it was updated successfully
 	return true, nil
-}
-
-func (r *UDPRouteReconciler) ensureParentsProgrammedCondition(
-	ctx context.Context,
-	udproute *gatewayv1alpha2.UDPRoute,
-	gateways []supportedGatewayWithCondition,
-	conditionStatus metav1.ConditionStatus,
-	conditionReason gatewayv1beta1.RouteConditionReason,
-	conditionMessage string,
-) (bool, error) {
-	// map the existing parentStatues to avoid duplications
-	parentStatuses := getParentStatuses(udproute, udproute.Status.Parents)
-
-	programmedCondition := metav1.Condition{
-		Type:               ConditionTypeProgrammed,
-		Status:             conditionStatus,
-		Reason:             string(conditionReason),
-		ObservedGeneration: udproute.Generation,
-		Message:            conditionMessage,
-		LastTransitionTime: metav1.Now(),
-	}
-	statusChanged := false
-	for _, g := range gateways {
-		gateway := g.gateway
-		parentRefKey := fmt.Sprintf("%s/%s", gateway.Namespace, gateway.Name)
-		parentStatus, ok := parentStatuses[parentRefKey]
-		if ok {
-			// update existing parent in status.
-			changed := setRouteParentStatusCondition(parentStatus, programmedCondition)
-			statusChanged = statusChanged || changed
-		} else {
-			// add a new parent if the parent is not found in status.
-			newParentStatus := &gatewayv1alpha2.RouteParentStatus{
-				ParentRef: gatewayv1alpha2.ParentReference{
-					Namespace:   lo.ToPtr(gatewayv1alpha2.Namespace(gateway.Namespace)),
-					Name:        gatewayv1alpha2.ObjectName(gateway.Name),
-					SectionName: lo.ToPtr(gatewayv1alpha2.SectionName(g.listenerName)),
-					// TODO: set port after gateway port matching implemented: https://github.com/Kong/kubernetes-ingress-controller/issues/3016
-				},
-				Conditions: []metav1.Condition{
-					programmedCondition,
-				},
-			}
-			udproute.Status.Parents = append(udproute.Status.Parents, *newParentStatus)
-			parentStatuses[parentRefKey] = newParentStatus
-			statusChanged = true
-		}
-	}
-
-	// update status if needed.
-	if statusChanged {
-		if err := r.Status().Update(ctx, udproute); err != nil {
-			return false, err
-		}
-		return true, nil
-	}
-	// no need to update if no status is changed.
-	return false, nil
 }